requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
lxml==4.9.1
multidict==6.0.2
psycopg2-binary==2.9.3
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
yarl==1.7.2
